refactor(confirm-delete-modal): relocate to components/ so components can consume it
ConfirmDeleteModal is a pure presentational, prop-driven modal with no screen
logic — it belongs in src/components/, not src/screens/. EditGoalView (a
component) needs to consume it to confirm step/sub-step deletions, but the
local/no-component-imports-screens ESLint boundary forbids components/ →
screens/ imports.

Move the modal to src/components/ConfirmDeleteModal/, fix its internal imports
(../../components/{Text,Card,Button} → ../{Text,Card,Button}), and update all
four screen consumers (BadgeDetail, EditMode, FocusMode, Goals) plus the a11y
contract test to the new path. Behavior unchanged — imports only.
