+++ apps/native-rd/src/components/ConfirmDeleteModal/ConfirmDeleteModal.styles.ts
