+++ apps/native-rd/src/components/ConfirmDeleteModal/index.ts
